Fix subscription
